requirements.txt
python-telegram-bot==13.7
feedparser
requests
